backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
opencv-python
numpy
imageio-ffmpeg
requests
azure-storage-blob
azure-data-tables
azure-storage-queue